[packaging] Disable udev_sync feature. Contributes to JB#36605
Disable udev_sync feature because it breaks booting. When it's enabled lvm
tool in initramfs stop to create /dev/sailfish/* devices, expecting that UDEV
will create it. But we have no UDEV in initramfs yet. So disable it for now.

But we need UDEV rules for systemd v225, so enable udev-rules.

Signed-off-by: Igor Zhbanov <igor.zhbanov@jolla.com>